move to dystopia folder, cleanup
