+++ src/dex/index.ts
@@ -41,8 +41,8 @@ import { KyberDmm } from './kyberdmm/kyberdmm';
 import { Platypus } from './platypus/platypus';
 import { GMX } from './gmx/gmx';
 import { WooFi } from './woo-fi/woo-fi';
-import { Dystopia } from './uniswap-v2/dystopia';
-import { DystopiaStable } from './uniswap-v2/dystopia-stable';
+import { Dystopia } from './uniswap-v2/dystopia/dystopia';
+import { DystopiaStable } from './uniswap-v2/dystopia/dystopia-stable';
 
 const LegacyDexes = [
   Curve,
