+++ src/dex/uniswap-v2/dystopia/dystopia-stable.ts
@@ -1,11 +1,11 @@
-import { UniswapV2, RESERVE_LIMIT } from './uniswap-v2';
-import { Network, NULL_ADDRESS } from '../../constants';
-import { Address, DexConfigMap, PoolLiquidity, Token } from '../../types';
-import { IDexHelper } from '../../dex-helper';
-import { DexParams } from './types';
-import { getDexKeysWithNetwork } from '../../utils';
-import dystopiaFactoryABI from '../../abi/uniswap-v2/DystFactory.json';
-import { BI_POWS } from '../../bigint-constants';
+import { UniswapV2, RESERVE_LIMIT } from '../uniswap-v2';
+import { Network, NULL_ADDRESS } from '../../../constants';
+import { Address, DexConfigMap, PoolLiquidity, Token } from '../../../types';
+import { IDexHelper } from '../../../dex-helper';
+import { DexParams } from '../types';
+import { getDexKeysWithNetwork } from '../../../utils';
+import dystopiaFactoryABI from '../../../abi/uniswap-v2/DystFactory.json';
+import { BI_POWS } from '../../../bigint-constants';
 import {
   DystopiaSharedPolygonConfig,
   getTopPoolsForTokenFiltered,
